Add 2-selection_sort.c

+++ 2-selection_sort.c
@@ -10,7 +10,7 @@
 void selection_sort(int *array, size_t size)
 {
     size_t i, j;
-    int tmp, min_index;
+    size_t tmp, min_index;
 
     if (array == NULL || size < 2)
         return;
